backend/djangoBackend/requirements.txt
Django==1.10.2
django-ssl-auth==1.0
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
django-cors-headers==1.2.2
pymysql
mysqlclient
uwsgi==2.0.14
paramiko >= 1.15.2
pysftp ==0.2.9